- API! T_scalar_type -> T_scalar; mat4::rotation() -> getRotation(); mat4::scaling() -> getScaling(); mat4::translation() -> getTranslation(); mat4::ortho() -> getOrtho(); etc. - API! Matrix4/3/2: removed [] operator; e(i,j) now specifies i-row, j-column; - Several enhancements and fixes to the matrix classes and others. - Matrix and quaternion regression tests.

+++ trunk/Visualization_Library_SDK/src/vl/GLSL.hpp
@@ -393,7 +393,7 @@ namespace vl
     bool getProgramBinary(GLenum& binary_format, std::vector<unsigned char>& binary) const;
 
     //! glProgramBinary wrapper: loads a program object with a program binary, see also http://www.opengl.org/sdk/docs/man4/xhtml/glProgramBinary.xml
-    bool programBinary(GLenum binary_format, const std::vector<unsigned char>& binary) { return programBinary(binary_format, &binary[0], binary.size()); }
+    bool programBinary(GLenum binary_format, const std::vector<unsigned char>& binary) { return programBinary(binary_format, &binary[0], (int)binary.size()); }
 
     //! glProgramBinary wrapper: loads a program object with a program binary, see also http://www.opengl.org/sdk/docs/man4/xhtml/glProgramBinary.xml
     bool programBinary(GLenum binary_format, const void* binary, int length);
